+++ src/sync/tests/variable_task.rs
@@ -1,8 +1,11 @@
-use crate::sync2::PollingTaskBuilder;
-use std::sync::atomic::AtomicU64;
-use std::sync::atomic::Ordering::SeqCst;
-use std::sync::{Arc, Mutex};
-use std::time::Duration;
+use crate::sync::PollingTaskBuilder;
+use std::{
+    sync::{
+        atomic::{AtomicU64, Ordering::SeqCst},
+        Arc, Mutex,
+    },
+    time::Duration,
+};
 
 #[tokio::test]
 async fn update_observed_on_next_poll_with_early_exit() {
@@ -11,11 +14,19 @@ async fn update_observed_on_next_poll_with_early_exit() {
     let (tx, rx) = tokio::sync::oneshot::channel();
     let tx = Mutex::new(Some(tx));
 
+    let counter_clone_2 = counter.clone();
+    let increases_on_second_call = move || {
+        if counter_clone_2.load(SeqCst) == 0 {
+            Duration::from_millis(0)
+        } else {
+            Duration::from_secs(5000)
+        }
+    };
+
     let _task = PollingTaskBuilder::new(Duration::from_millis(0))
         .wait_for_clean_exit(None)
-        .self_updating_task(move |interval: &mut Duration| {
+        .variable_task(increases_on_second_call, move || {
             counter_clone.fetch_add(1, SeqCst);
-            *interval = Duration::from_secs(5000);
             if let Some(tx) = tx.lock().unwrap().take() {
                 tx.send(true).unwrap();
             }
@@ -35,19 +46,20 @@ async fn slow_poll_exits_early() {
     {
         let _task = PollingTaskBuilder::new(Duration::from_millis(0))
             .wait_for_clean_exit(None)
-            .self_updating_task_with_checker(move |interval: &mut Duration, checker: &dyn Fn() -> bool| {
-                tx.lock().unwrap().take().unwrap().send(true).unwrap();
+            .variable_task_with_checker(
+                || Duration::from_secs(5000),
+                move |checker: &dyn Fn() -> bool| {
+                    tx.lock().unwrap().take().unwrap().send(true).unwrap();
 
-                loop {
-                    if !checker() {
-                        break;
+                    loop {
+                        if !checker() {
+                            break;
+                        }
                     }
-                }
 
-                // Prevent issues caused by cycling a second time.
-                *interval = Duration::from_secs(5000);
-                tx_exit.lock().unwrap().take().unwrap().send(true).unwrap();
-            });
+                    tx_exit.lock().unwrap().take().unwrap().send(true).unwrap();
+                },
+            );
 
         // Guarantee we polled at least once
         rx.await.unwrap();
